requirements.txt
telethon
python-dotenv
colorama
tqdm
